refactor: rename erroClasses file to errorClasses
